map ref type correctly
