+++ lib/loaders/file.js
@@ -1,6 +1,5 @@
 const fs = require('fs')
 const path = require('path')
-const jsyaml = require('js-yaml')
 const { getRefFilePath } = require('../utils')
 
 var cwd = process.cwd()
@@ -13,7 +12,7 @@ var cwd = process.cwd()
  * @returns {*}
  * @private
  */
-module.exports = function (refValue, options) {
+module.exports = function(refValue, options) {
   let refPath = refValue
   const baseFolder = options.baseFolder ? path.resolve(cwd, options.baseFolder) : cwd
 
@@ -24,31 +23,12 @@ module.exports = function (refValue, options) {
   }
 
   const filePath = getRefFilePath(refPath)
-  const filePathLowerCase = filePath.toLowerCase()
 
   let newValue
   try {
     var data = fs.readFileSync(filePath, 'utf8')
-    if (filePathLowerCase.endsWith('.json')) {
-      newValue = parseJson(data) || parseYaml(data)
-    } else if (filePathLowerCase.endsWith('.yml') || filePathLowerCase.endsWith('.yaml')) {
-      newValue = parseYaml(data)
-    } else {
-      newValue = parseJson(data) || parseYaml(data)
-    }
+    newValue = JSON.parse(data)
   } catch (e) { }
 
   return newValue
 };
-
-function parseJson(data) {
-  try {
-    return JSON.parse(data)
-  } catch (e) {}
-}
-
-function parseYaml(data) {
-  try {
-    return jsyaml.load(data)
-  } catch (e) {}
-}
