+++ lib/utils.js
@@ -18,7 +18,6 @@ function getRefValue (ref) {
 }
 exports.getRefValue = getRefValue
 
-const yamlSuffixRegex = /\.ya?ml/ // regex for matching .yaml/.yml files
 /**
  * Gets the type of $ref from search result object.
  * @param ref The search result object from prop-search or a ref object
@@ -34,20 +33,12 @@ function getRefType (ref) {
     if (validUrl.isWebUri(val)) {
       return 'web'
     }
-    
-    const varLowerCase = val.toLowerCase()
-    if (
-      (validUrl.isUri(varLowerCase) && varLowerCase.startsWith('file:')) ||
-      varLowerCase.includes('.json') ||
-      varLowerCase.includes('.yml') ||
-      varLowerCase.includes('.yaml') ||
-      isValidPath(varLowerCase.split['#'])
-    ) {
-      return 'file'
-    }
-    if ((val.charAt(0) === '#') || (!validUrl.isUri(val))) {
+
+    if ((val.charAt(0) === '#')) {
       return 'local'
     }
+    
+    return 'file'
   }
 }
 exports.getRefType = getRefType
